Added a request of user's permission for use of location
If the user uses an ad block that blocks Ip-Address-Locating (for privacy protection) then the app's first location doesn't load and throws an error (blocked by client).

I added a request to use user's location, as well as a note to disable ad block if they want it to load properly.

I also updated error handling so that it wouldn't throw (city not found) error if the previous mentioned error happened.